Replace columns in the rowData

The function replaces columns in the rowData of assays from a QFeatures object. This function can also be used to remove column in the rowData (see examples).

replaceRowDataCols(object, i, col, value)

Arguments

object
An object of class QFeatures

i
A numeric() or character() vector indicating from which assays the rowData should be taken.

col
A numeric() or character() vector indicating the column(s) to replace.

value
A list() of same length as i. Each element should contain the replacement values to insert in the rowData. If length(col) > 1, each element should be a matrix with the same rows as its corresponding assay and the same number of columns as the length of col.

Value

An object of class QFeatures

Examples

data("feat2")
rowData(feat2)[[1]]

#> DataFrame with 10 rows and 2 columns
#>          Prot         x
#>   <character> <numeric>
#> a          Pa  2.067646
#> b          Pb -0.188981
#> c          Pc  0.266870
#> d          Pd  1.671331
#> e          Pe -1.857170
#> f          Pf  1.166811
#> g          Pg  0.316521
#> h          Ph  0.976154
#> i          Pi  0.117673
#> j          Pj  2.420734

## Add a new column to all assay
values <- lapply(rowData(feat2), function(x) rep("bar", nrow(x)))
feat2 <- replaceRowDataCols(feat2, 1:3, "foo", values)
rowData(feat2)[[1]]

#> DataFrame with 10 rows and 3 columns
#>          Prot         x         foo
#>   <character> <numeric> <character>
#> a          Pa  2.067646         bar
#> b          Pb -0.188981         bar
#> c          Pc  0.266870         bar
#> d          Pd  1.671331         bar
#> e          Pe -1.857170         bar
#> f          Pf  1.166811         bar
#> g          Pg  0.316521         bar
#> h          Ph  0.976154         bar
#> i          Pi  0.117673         bar
#> j          Pj  2.420734         bar

## Update column
values <- lapply(rowData(feat2), function(x) rep("barNew", nrow(x)))
feat2 <- replaceRowDataCols(feat2, 1:3, "foo", values)
rowData(feat2)[[1]]

#> DataFrame with 10 rows and 3 columns
#>          Prot         x         foo
#>   <character> <numeric> <character>
#> a          Pa  2.067646      barNew
#> b          Pb -0.188981      barNew
#> c          Pc  0.266870      barNew
#> d          Pd  1.671331      barNew
#> e          Pe -1.857170      barNew
#> f          Pf  1.166811      barNew
#> g          Pg  0.316521      barNew
#> h          Ph  0.976154      barNew
#> i          Pi  0.117673      barNew
#> j          Pj  2.420734      barNew

## Remove column
values <- lapply(1:3, function(x) NULL)
feat2 <- replaceRowDataCols(feat2, 1:3, "foo", values)
rowData(feat2)[[1]]

#> DataFrame with 10 rows and 2 columns
#>          Prot         x
#>   <character> <numeric>
#> a          Pa  2.067646
#> b          Pb -0.188981
#> c          Pc  0.266870
#> d          Pd  1.671331
#> e          Pe -1.857170
#> f          Pf  1.166811
#> g          Pg  0.316521
#> h          Ph  0.976154
#> i          Pi  0.117673
#> j          Pj  2.420734